+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 14-08-23
+
+## Risk Engine Tests
+- Binary Commands: Symbols, Accounts
+- PlaceOrder
+- User Management
+- Reset
+- PostProcess
+
 ## 13-08-23
 
 - поправить тесты